Updated Create, Edit, and Delete pages and customized the Details page.

+++ Mvc/Controllers/StudentsController.cs
@@ -1,3 +1,4 @@
+// controls the aspects of the Student tab of the web app
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -37,7 +38,19 @@ namespace Mvc.Controllers
             }
 
             var student = await _context.Students
+                // Include and ThenInclude methods cause the context to load the
+                // Student.Enrollments navigation property, and within each enrollment
+                // the Enrollment.Course navigation property
+                .Include(s => s.Enrollments)
+                    .ThenInclude(e => e.Course)
+
+                // AsNoTracking method improves performance in scenarios where the
+                // entities returned won't be updated in the current context's lifetime
+                .AsNoTracking()
+
+                // FirstOrDefaultAsync method to retrieve a single Student entity
                 .FirstOrDefaultAsync(m => m.ID == id);
+
             if (student == null)
             {
                 return NotFound();
@@ -47,6 +60,7 @@ namespace Mvc.Controllers
         }
 
         // GET: Students/Create
+        // displays view so that user can choose to create
         public IActionResult Create()
         {
             return View();
@@ -55,15 +69,29 @@ namespace Mvc.Controllers
         // POST: Students/Create
         // To protect from overposting attacks, enable the specific properties you want to bind to.
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+
+        // adds the Student entity created by the ASP.NET Core MVC model binder to the
+        // Students entity set and then saves the changes to the database
         [HttpPost]
-        [ValidateAntiForgeryToken]
+        [ValidateAntiForgeryToken] // helps prevent cross-site request forgery
         public async Task<IActionResult> Create([Bind("ID,LastName,FirstMidName,EnrollmentDate")] Student student)
         {
-            if (ModelState.IsValid)
+            try
             {
-                _context.Add(student);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                // model validation check
+                if (ModelState.IsValid)
+                {
+                    _context.Add(student);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+            }
+            catch (DbUpdateException /* ex */)
+            {
+                // log the error (uncomment ex variable name and write a log)
+                ModelState.AddModelError("", "Unable to save changes. " +
+                    "Try again, and if the problem persists " +
+                    "see your system administrator.");
             }
             return View(student);
         }
@@ -87,40 +115,47 @@ namespace Mvc.Controllers
         // POST: Students/Edit/5
         // To protect from overposting attacks, enable the specific properties you want to bind to.
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
+        // allows user to update certain attributes about the Student
+        [HttpPost, ActionName("Edit")] // uses action method to connect to Edit method
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ID,LastName,FirstMidName,EnrollmentDate")] Student student)
+        public async Task<IActionResult> EditPost(int? id)
         {
-            if (id != student.ID)
+            if (id == null)
             {
                 return NotFound();
             }
 
-            if (ModelState.IsValid)
+            // reads the existing entity
+            var studentToUpdate = await _context.Students.FirstOrDefaultAsync(s => s.ID == id);
+            // update fields in the retrieved entity based on user input in the posted form data
+            if (await TryUpdateModelAsync<Student>(
+                // updateable attributes
+                studentToUpdate,
+                "",
+                s => s.FirstMidName, s => s.LastName, s => s.EnrollmentDate))
             {
                 try
                 {
-                    _context.Update(student);
+                    // SaveChanges -> Entity Framework creates SQL statements to update the database row
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (DbUpdateException /* ex */)
                 {
-                    if (!StudentExists(student.ID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    //Log the error (uncomment ex variable name and write a log.)
+                    ModelState.AddModelError("", "Unable to save changes. " +
+                        "Try again, and if the problem persists, " +
+                        "see your system administrator.");
                 }
-                return RedirectToAction(nameof(Index));
             }
-            return View(student);
+            return View(studentToUpdate);
         }
 
         // GET: Students/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        // displays view so that user can choose to delete
+        // manages error reporting
+        // accepts an optional parameter that indicates whether the method was called after a failure to save changes
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -128,24 +163,49 @@ namespace Mvc.Controllers
             }
 
             var student = await _context.Students
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.ID == id);
             if (student == null)
             {
                 return NotFound();
             }
 
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                    "Delete failed. Try again, and if the problem persists " +
+                    "see your system administrator.";
+            }
+
             return View(student);
         }
 
         // POST: Students/Delete/5
+        // deletes selected entity
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            // retrieves selected entity
             var student = await _context.Students.FindAsync(id);
-            _context.Students.Remove(student);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if(student == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
+            try
+            {
+                // calls Remove method to set entity's setting to Deleted
+                _context.Students.Remove(student);
+                // SQL DELETE command is generated
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException /* ex */)
+            {
+                // log the error (uncomment ex variable name and write a log)
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
         }
 
         private bool StudentExists(int id)
